refactor: rearrange gene map modules
