+++ packages_rs/nextclade-cli/src/bin/generate_jsonschema.rs
@@ -4,11 +4,11 @@ use eyre::Report;
 use log::LevelFilter;
 use nextclade::analyze::pcr_primers::PcrPrimer;
 use nextclade::analyze::virus_properties::{PhenotypeAttrDesc, VirusProperties};
+use nextclade::gene::gene_map::GeneMap;
 use nextclade::io::dataset::{DatasetTagJson, DatasetsIndexJson};
 use nextclade::io::errors_csv::ErrorsFromWeb;
 use nextclade::io::fasta::FastaRecord;
 use nextclade::io::file::create_file_or_stdout;
-use nextclade::io::gene_map::GeneMap;
 use nextclade::io::json::json_write_impl;
 use nextclade::io::nextclade_csv::CsvColumnConfig;
 use nextclade::qc::qc_config::QcConfig;
