+++ packages_rs/nextclade-cli/src/cli/nextalign_loop.rs
@@ -7,9 +7,9 @@ use log::info;
 use nextclade::align::gap_open::{get_gap_open_close_scores_codon_aware, get_gap_open_close_scores_flat};
 use nextclade::align::params::AlignPairwiseParams;
 use nextclade::alphabet::nuc::{to_nuc_seq, to_nuc_seq_replacing};
+use nextclade::gene::gene_map::{filter_gene_map, GeneMap};
+use nextclade::gene::gene_map_display::gene_map_to_table_string;
 use nextclade::io::fasta::{read_one_fasta, FastaReader, FastaRecord};
-use nextclade::io::gene_map::gene_map_to_string;
-use nextclade::io::gene_map::{filter_gene_map, GeneMap};
 use nextclade::run::nextalign_run_one::nextalign_run_one;
 use nextclade::translate::translate_genes_ref::translate_genes_ref;
 use nextclade::types::outputs::NextalignOutputs;
@@ -66,7 +66,7 @@ pub fn nextalign_run(run_args: NextalignRunArgs) -> Result<(), Report> {
     None => GeneMap::new(),
   };
 
-  info!("Gene map:\n{}", gene_map_to_string(&gene_map)?);
+  info!("Gene map:\n{}", gene_map_to_table_string(&gene_map)?);
 
   let gap_open_close_nuc = &get_gap_open_close_scores_codon_aware(ref_seq, &gene_map, &alignment_params);
   let gap_open_close_aa = &get_gap_open_close_scores_flat(ref_seq, &alignment_params);
