+++ packages_rs/nextclade-cli/src/cli/nextalign_ordered_writer.rs
@@ -2,9 +2,9 @@ use crate::cli::nextalign_loop::NextalignRecord;
 use eyre::{Report, WrapErr};
 use log::{info, warn};
 use nextclade::alphabet::nuc::from_nuc_seq;
+use nextclade::gene::gene_map::GeneMap;
 use nextclade::io::errors_csv::ErrorsCsvWriter;
 use nextclade::io::fasta::{FastaPeptideWriter, FastaRecord, FastaWriter};
-use nextclade::io::gene_map::GeneMap;
 use nextclade::io::insertions_csv::InsertionsCsvWriter;
 use nextclade::translate::translate_genes::Translation;
 use nextclade::types::outputs::NextalignOutputs;
