+++ packages_rs/nextclade-cli/src/cli/nextclade_loop.rs
@@ -13,9 +13,9 @@ use nextclade::align::params::AlignPairwiseParams;
 use nextclade::alphabet::nuc::{to_nuc_seq, to_nuc_seq_replacing, Nuc};
 use nextclade::analyze::find_aa_motifs::find_aa_motifs;
 use nextclade::analyze::phenotype::get_phenotype_attr_descs;
+use nextclade::gene::gene_map_display::gene_map_to_table_string;
 use nextclade::io::fasta::{FastaReader, FastaRecord};
 use nextclade::io::fs::has_extension;
-use nextclade::io::gene_map::gene_map_to_string;
 use nextclade::io::json::json_write;
 use nextclade::io::nextclade_csv::CsvColumnConfig;
 use nextclade::make_error;
@@ -129,7 +129,7 @@ pub fn nextclade_run(run_args: NextcladeRunArgs) -> Result<(), Report> {
   alignment_params.merge_opt(run_args.alignment_params);
 
   info!("Alignment parameters (final):\n{alignment_params:#?}");
-  info!("Gene map:\n{}", gene_map_to_string(gene_map)?);
+  info!("Gene map:\n{}", gene_map_to_table_string(gene_map)?);
 
   let gap_open_close_nuc = &get_gap_open_close_scores_codon_aware(ref_seq, gene_map, &alignment_params);
   let gap_open_close_aa = &get_gap_open_close_scores_flat(ref_seq, &alignment_params);
