+++ packages_rs/nextclade-cli/src/cli/nextclade_ordered_writer.rs
@@ -4,9 +4,9 @@ use itertools::Itertools;
 use log::{info, warn};
 use nextclade::alphabet::nuc::from_nuc_seq;
 use nextclade::analyze::virus_properties::PhenotypeAttrDesc;
+use nextclade::gene::gene_map::GeneMap;
 use nextclade::io::errors_csv::ErrorsCsvWriter;
 use nextclade::io::fasta::{FastaPeptideWriter, FastaRecord, FastaWriter};
-use nextclade::io::gene_map::GeneMap;
 use nextclade::io::insertions_csv::InsertionsCsvWriter;
 use nextclade::io::ndjson::NdjsonFileWriter;
 use nextclade::io::nextclade_csv::{CsvColumnConfig, NextcladeResultsCsvFileWriter};
