+++ packages_rs/nextclade-cli/src/dataset/dataset_download.rs
@@ -6,10 +6,10 @@ use itertools::Itertools;
 use log::LevelFilter;
 use nextclade::analyze::pcr_primers::PcrPrimer;
 use nextclade::analyze::virus_properties::VirusProperties;
+use nextclade::gene::gene_map::{filter_gene_map, GeneMap};
 use nextclade::io::dataset::{Dataset, DatasetsIndexJson};
 use nextclade::io::fasta::{read_one_fasta, read_one_fasta_str, FastaRecord};
 use nextclade::io::fs::absolute_path;
-use nextclade::io::gene_map::{filter_gene_map, GeneMap};
 use nextclade::io::json::json_parse_bytes;
 use nextclade::make_error;
 use nextclade::qc::qc_config::QcConfig;
