+++ packages_rs/nextclade-web/src/wasm/analyze.rs
@@ -9,8 +9,8 @@ use nextclade::analyze::pcr_primers::PcrPrimer;
 use nextclade::analyze::phenotype::get_phenotype_attr_descs;
 use nextclade::analyze::virus_properties::{AaMotifsDesc, PhenotypeAttrDesc, VirusProperties};
 use nextclade::gene::gene::Gene;
+use nextclade::gene::gene_map::GeneMap;
 use nextclade::io::fasta::read_one_fasta_str;
-use nextclade::io::gene_map::GeneMap;
 use nextclade::io::json::json_stringify;
 use nextclade::io::nextclade_csv::CsvColumnConfig;
 use nextclade::qc::qc_config::QcConfig;
