+++ packages_rs/nextclade-web/src/wasm/main.rs
@@ -3,9 +3,9 @@ use eyre::{Report, WrapErr};
 use itertools::Itertools;
 use nextclade::analyze::pcr_primers::PcrPrimer;
 use nextclade::analyze::virus_properties::{AaMotifsDesc, PhenotypeAttrDesc, VirusProperties};
+use nextclade::gene::gene_map::GeneMap;
 use nextclade::io::errors_csv::{errors_to_csv_string, ErrorsFromWeb};
 use nextclade::io::fasta::{read_one_fasta_str, FastaReader, FastaRecord};
-use nextclade::io::gene_map::GeneMap;
 use nextclade::io::insertions_csv::insertions_to_csv_string;
 use nextclade::io::json::{json_parse, json_stringify};
 use nextclade::io::nextclade_csv::{results_to_csv_string, CsvColumnConfig};
