+++ packages_rs/nextclade/src/align/backtrace.rs
@@ -103,7 +103,7 @@ mod tests {
   use crate::align::params::AlignPairwiseParams;
   use crate::align::score_matrix;
   use crate::alphabet::nuc::{to_nuc_seq, Nuc};
-  use crate::io::gene_map::GeneMap;
+  use crate::gene::gene_map::GeneMap;
   use eyre::Report;
   use pretty_assertions::assert_eq;
   use rstest::{fixture, rstest};
