+++ packages_rs/nextclade/src/align/gap_open.rs
@@ -2,7 +2,7 @@ use crate::align::params::AlignPairwiseParams;
 use crate::alphabet::nuc::Nuc;
 use crate::coord::range::NucRefGlobalRange;
 use crate::gene::gene::GeneStrand;
-use crate::io::gene_map::GeneMap;
+use crate::gene::gene_map::GeneMap;
 use either::Either;
 
 pub type GapScoreMap = Vec<i32>;
