+++ packages_rs/nextclade/src/features/mod.rs
@@ -3,5 +3,4 @@ pub mod feature_group;
 pub mod feature_tree;
 pub mod feature_tree_format;
 pub mod feature_type;
-pub mod gene_map;
 pub mod sequence_region;
