+++ packages_rs/nextclade/src/gene/mod.rs
@@ -2,6 +2,8 @@ pub mod cds;
 pub mod cds_segment;
 pub mod frame;
 pub mod gene;
+pub mod gene_map;
+pub mod gene_map_display;
 pub mod genotype;
 pub mod phase;
 pub mod protein;
