+++ packages_rs/nextclade/src/io/fasta.rs
@@ -1,9 +1,9 @@
 use crate::alphabet::aa::from_aa_seq;
 use crate::constants::REVERSE_COMPLEMENT_SUFFIX;
+use crate::gene::gene_map::GeneMap;
 use crate::io::compression::Decompressor;
 use crate::io::concat::concat;
 use crate::io::file::{create_file_or_stdout, open_file_or_stdin, open_stdin};
-use crate::io::gene_map::GeneMap;
 use crate::translate::translate_genes::CdsTranslation;
 use crate::{make_error, make_internal_error};
 use eyre::{Report, WrapErr};
