+++ packages_rs/nextclade/src/translate/aa_alignment_ranges.rs
@@ -1,7 +1,7 @@
 use crate::coord::coord_map::local_to_codon_range_exclusive;
 use crate::coord::position::PositionLike;
 use crate::coord::range::{intersect, AaRefRange, NucRefGlobalRange, NucRefLocalRange};
-use crate::io::gene_map::GeneMap;
+use crate::gene::gene_map::GeneMap;
 use crate::translate::translate_genes::Translation;
 use eyre::Report;
 use itertools::Itertools;
